fix tests failing as CARGO_MANIFEST_DIR in path
The tests run in a subdirectory of the source directory. Fish detects
that we are running from a subdirectory of the CARGO_MANIFEST_DIR, and
that causes fish to try and find the data files relative to
CARGO_MANIFEST_DIR, which no longer exists.

Move the source directory into its own subdirectory, so that fish won't
confuse it.