Run tests with fish's upstream tests
Drop the "tests" from our feedstock and run the tests from upstream fish
instead, so that we don't need to maintain a fork of the tests in our
feedstock.

We created our fork of "tests" because we needed to run a bunch of tests
that were present in the fish development branch. Those tests are now
part of the fish 4.1.0 release, so we no longer need a fork.